WIP: DirectVariableAccess

+++ src/tint/writer/glsl/generator_impl_member_accessor_test.cc
@@ -418,7 +418,7 @@ void tint_symbol() {
   int a = 2;
   int b = 4;
   int c = 3;
-  int x = data.a[((a + b) - c)];
+  int x = data.a[uint(((a + b) - c))];
 }
 
 void main() {
