+++ test/tint/bug/fxc/dyn_array_idx/read/workgroup.wgsl.expected.glsl
@@ -24,7 +24,7 @@ void f(uint local_invocation_index) {
     }
   }
   barrier();
-  result.tint_symbol = s.data[ubo.dynamic_idx];
+  result.tint_symbol = s.data[uint(ubo.dynamic_idx)];
 }
 
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
